+++ tests/disko_lib/types_disk/partial_failure_dos_table-generate-result.json
