+++ tests/disko_lib/types_disk/test_types_disk.py
@@ -10,7 +10,7 @@ CURRENT_DIR = Path(__file__).parent
 
 
 def test_generate_config_partial_failure_dos_table() -> None:
-    with open(CURRENT_DIR / "lsblk-output.json") as f:
+    with open(CURRENT_DIR / "partial_failure_dos_table-lsblk-output.json") as f:
         lsblk_result = device.list_block_devices(f.read())
 
     assert isinstance(lsblk_result, DiskoSuccess)
@@ -26,7 +26,7 @@ def test_generate_config_partial_failure_dos_table() -> None:
     }
 
     assert result.messages[1].factory == warn_generate_partial_failure
-    with open(CURRENT_DIR / "generate-result.json") as f:
+    with open(CURRENT_DIR / "partial_failure_dos_table-generate-result.json") as f:
         assert result.messages[1].details["partial_config"] == json.load(f)
     assert result.messages[1].details["failed_devices"] == [PosixPath("/dev/sdc")]
     assert result.messages[1].details["successful_devices"] == [
